market-dashboard/requirements.txt
streamlit>=1.30
plotly>=5.20
pandas>=2.2
numpy>=1.26
yfinance>=0.2.52

python-dotenv~=1.1.1
pytest~=8.4.1
ta~=0.11.0